+++ CHANGELOG.md
@@ -6,9 +6,9 @@
 
 #### Dependencies
 
-- Updated eslint from version 9.6.0 to 9.9.0
+- Updated eslint from version 9.6.0 to 9.9.1
 - Updated eslint-plugin-jsdoc from version 48.5.0 to 50.2.2
-- Updated gulp-eslint from version 2.1.0 to 2.2.0
+- Updated gulp-eslint-new from version 2.1.0 to 2.3.0
 
 ---